wip for evaluating on test state
