+++ test.py
@@ -28,51 +28,51 @@ import numpy as np
 from utils import HotWheelsStates
 
 
-def make_retro(
-    *,
-    game,
-    state: HotWheelsStates = HotWheelsStates.DEFAULT,
-    max_episode_steps=4500,
-    render_mode="human",
-    **kwargs,
-):
-    print(f"Using ", f"state: {state}.state", f"info: {state}.json", sep="\n")
-    env = retro.make(
-        game,
-        state=f"{state}.state",
-        info=retro.data.get_file_path(
-            "HotWheelsStuntTrackChallenge-gba", f"{state}.json"
-        ),
-        render_mode=render_mode,
-        **kwargs,
-    )
-    env = Monitor(env)
-    env = StochasticFrameSkip(env, n=4, stickprob=0.25)
-    env = TerminateOnCrash(env)
-    env = PenalizeHittingWalls(env)
-    env = HotWheelsDiscretizer(env)
+# def make_retro(
+#     *,
+#     game,
+#     state: HotWheelsStates = HotWheelsStates.DEFAULT,
+#     max_episode_steps=4500,
+#     render_mode="human",
+#     **kwargs,
+# ):
+#     print(f"Using ", f"state: {state}.state", f"info: {state}.json", sep="\n")
+#     env = retro.make(
+#         game,
+#         state=f"{state}.state",
+#         info=retro.data.get_file_path(
+#             "HotWheelsStuntTrackChallenge-gba", f"{state}.json"
+#         ),
+#         render_mode=render_mode,
+#         **kwargs,
+#     )
+#     env = Monitor(env)
+#     env = StochasticFrameSkip(env, n=4, stickprob=0.25)
+#     env = TerminateOnCrash(env)
+#     env = PenalizeHittingWalls(env)
+#     env = HotWheelsDiscretizer(env)
 
-    return env
-
-
-def wrap_deepmind_retro(env):
-    """
-    Configure environment for retro games, using config similar to DeepMind-style Atari in openai/baseline's wrap_deepmind
-    """
-    env = WarpFrame(env)
-    env = ClipRewardEnv(env)
-    return env
+#     return env
 
 
-def make_env():
-    env = make_retro(
-        game="HotWheelsStuntTrackChallenge-gba",
-        state=HotWheelsStates.DINO_BONEYARD_MULTI,
-        scenario=None,
-    )
-    env = wrap_deepmind_retro(env)
-    env = Viewer(env)
-    return env
+# def wrap_deepmind_retro(env):
+#     """
+#     Configure environment for retro games, using config similar to DeepMind-style Atari in openai/baseline's wrap_deepmind
+#     """
+#     env = WarpFrame(env)
+#     env = ClipRewardEnv(env)
+#     return env
+
+
+# def make_env():
+#     env = make_retro(
+#         game="HotWheelsStuntTrackChallenge-gba",
+#         state=HotWheelsStates.DINO_BONEYARD_MULTI,
+#         scenario=None,
+#     )
+#     env = wrap_deepmind_retro(env)
+#     env = Viewer(env)
+#     return env
 
 
 # venv = VecTransposeImage(VecFrameStack(DummyVecEnv([make_env] * 1), n_stack=4))
@@ -108,3 +108,90 @@ def make_env():
 
 # finally:
 #     venv.close()
+
+# import time
+
+# # create eval env
+# state = HotWheelsStates.TREX_VALLEY_SINGLE
+# env = retro.make(
+#     "HotWheelsStuntTrackChallenge-gba",
+#     render_mode="human",
+#     state=f"{state}.state",
+#     info=retro.data.get_file_path(
+#         "HotWheelsStuntTrackChallenge-gba", f"{state}.json"
+#     ),
+# )
+
+# full_track_state = env.unwrapped.em.get_state()
+
+# env.close()
+# del env
+
+# # create training env stae
+# env = retro.make(
+#     "HotWheelsStuntTrackChallenge-gba",
+#     render_mode="human",
+#     state=f"{232}.state",
+#     info=retro.data.get_file_path(
+#         "HotWheelsStuntTrackChallenge-gba", f"{state}.json"
+#     ),
+# )
+
+# obs, info = env.reset(seed=42)
+
+# n_step = 0
+# while True:
+#     n_step += 1
+#     print(n_step, "232", info)
+    
+#     if n_step >= 1000:
+#         break
+
+#     act = env.action_space.sample()
+
+#     obs, reward, term, trun, info = env.step(act)
+#     if term or trun:
+#         env.close()
+#         break
+
+# print("loading new state")
+
+# #env.em.set_state(full_track_state)
+# env.unwrapped.load_state(f"{HotWheelsStates.TREX_VALLEY_SINGLE}.state")
+# env.unwrapped.data.reset()
+# env.unwrapped.data.update_ram()
+
+# obs, info = env.reset(seed=42)
+
+
+# n_step = 0
+# while True:
+#     n_step += 1
+#     print(n_step, "full", info)
+
+#     if n_step >= 1000:
+#         break
+
+#     act = env.action_space.sample()
+
+#     obs, reward, term, trun, info = env.step(act)
+#     if term or trun:
+#         print(info, reward)
+#         env.close()
+#         break
+
+#     if n_step == 1:
+#         time.sleep(5)
+
+
+
+
+
+
+"""
+
+
+
+"""
+
+
